fix: detect quota replica changes across int32 wrap boundaries
Even with exact resource-usage calculation, the admission prefilter summed old and new replica counts in int32. A change of2^32 replicas compared equal and skipped quota evaluation entirely; scale-down changes across the same boundary could also fail to release reservations.

Compare totals accumulated in int64 so only genuinely equal scheduled replica counts take the unchanged fast path. Preserve existing component and resource-request change detection.

New positive/negative2^32 change cases and a full update-admission regression fail before this fix. The update is now denied with exact4294967296m requested CPU, while equal large totals remain unchanged. Complete binding-webhook/helper/federated-quota race suites and pinned lint pass.

+++ pkg/webhook/resourcebinding/validating.go
@@ -424,13 +424,13 @@ func isResourceRequestChanged(oldRB, newRB *workv1alpha2.ResourceBinding) bool {
 }
 
 func isScheduledReplicasChanged(oldRB, newRB *workv1alpha2.ResourceBinding) bool {
-	oldScheduledReplicas := int32(0)
+	oldScheduledReplicas := int64(0)
 	for _, c := range oldRB.Spec.Clusters {
-		oldScheduledReplicas += c.Replicas
+		oldScheduledReplicas += int64(c.Replicas)
 	}
-	newScheduledReplicas := int32(0)
+	newScheduledReplicas := int64(0)
 	for _, c := range newRB.Spec.Clusters {
-		newScheduledReplicas += c.Replicas
+		newScheduledReplicas += int64(c.Replicas)
 	}
 	return oldScheduledReplicas != newScheduledReplicas
 }
